refactor(project): restructure to Go standard layout
- Updated README.md build command to target ./cmd/dab-downloader
- Added detailed project architecture section reflecting new structure
- Added golang.org/x/time dependency in go.mod and updated go.sum
- Moved from flat file organization to cmd/ and internal/ packages for better modularity
